Moved Navbar component and stylesheet into own directory with index.js export file
